+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/drizzle
 
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.78.0
+
 ## 0.3.5
 
 ### Patch Changes